Extract vertical moves functionality into module
We aren't using inheritance because we don't want our Rook to know, how
to move diagonally. But with all functionality (forward, backward, etc.)
in a single module, we're doing the exactly same thing.

We could've created a Rook Object and call #upper_moves_right on it.
That's the reason we're isolating different behaviors into different
modules, so Rook doesn't know how to move diagonally and Bishop doesn't
know how to move vertically