Print ranks from 1~8
Earlier I was printing ranks from 8~1 and this wasn't intuitive. It
resulted in a bug wherein update() method was taking rank and
converting it into row_index. Because of this bug our markers for
possible positions were printing incorrectly.

Now it's intuitive and easier to manipulate as row_index corrrespond to
rank.