Fixbug: Piece need not to keep track of Piece
Yes Piece will need to look at Board to calculate it's possible moves.
But we can pass that context as a method parameter. This way not every
piece will have to keep track of state of the board.

The piece which we'll want to move, will be supplied 'board' as a method
parameter. This way other pieces doesn't care what the State of the
board is.