Display markers and captures
We display markers and captures when a piece is selected. So the given
functionality should be placed in Piece class. Because it's like we're
asking a piece (sending message), i.e. what are the possible actions
you can perform?

Piece responds this by painting markers to squares where it can move &
changing color of square to PURPLE which it can capture.