Model piece belongs to board
A piece belongs to Board but it's piece who should track, which board it
belongs to. I didn't model this relationship earlier because I thought
whichever piece is selected, only that piece has to know the state of
board. To calculate it's possible moves. Rest of the piece doesn't need
to keep track of Board state at all times.

But sooner I found that in my Piece class, every method had 'board' as a
parameter. So I let the piece keep track of board.